pyproject.toml
[project]
name = "npo"
version = "0.1.0"
description = "REST API to manage HD image"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiosqlite>=0.22.0",
  "alembic>=1.17.2",
  "fastapi[standard]>=0.125.0",
  "greenlet>=3.3.0",
  "pydantic-settings>=2.12.0",
  "pyexiftool>=0.5.6",
  "python-dotenv>=1.2.1",
  "python-multipart>=0.0.21",
  "pyvips>=3.1.1",
  "sqlalchemy>=2.0.45",
  "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "pylint>=4.0.4",
]

[tool.setuptools]
  package-dir = {"" = "src"}

[tool.setuptools.packages.find]
  where = ["src"]
  namespaces = true

[tool.black]
line-length = 100
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.tox
    | \.venv
  )/
)
'''

[tool.isort]
profile = "black"

[tool.pylint.main]
ignored-modules = ["alembic.context", "alembic.op"]

[tool.pylint.format]
max-line-length = "100"

[tool.pylint.messages_control]
disable = [
    "duplicate-code",
    "too-few-public-methods",
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
]

[tool.pylint.master]
init-hook = "import sys; sys.path.insert(0, 'src')"

[tool.alembic]
script_location = "%(here)s/migrations"
prepend_sys_path = [
    ".",
    "src",
]

